Staging

+++ src/admin/services/properties.service.mjs
@@ -1,11 +1,12 @@
 import { LeaseAggrements } from "../../user/models/leaseAggrements.model.mjs";
 import { Property } from "../../user/models/property.model.mjs";
 import { ObjectId } from 'bson';
-
+import { Inspection } from "../../user/models/inspection.model.mjs";
+import { InspectionStatus } from "../../user/enums/inspection.enums.mjs";
 
 async function leaseAggrementsList(filters) {
-  if(filters) {
-    const data = await LeaseAggrements.find({uploadedBy : filters})
+  if (filters) {
+    const data = await LeaseAggrements.find({ uploadedBy: filters })
     return {
       data: data,
       message: `successfully fetched  list`,
@@ -21,9 +22,8 @@ async function leaseAggrementsList(filters) {
       statusCode: 201,
     };
   }
-  
-}
 
+}
 
 async function getPropertiesList() {
   const data = await Property.aggregate([{
@@ -86,7 +86,6 @@ async function getPropertiesList() {
   };
 }
 
-
 async function getPropertyByID(id) {
   const id1 = new ObjectId(id)
   const data = await Property.aggregate([
@@ -127,14 +126,44 @@ async function getPropertyByID(id) {
 }
 
 async function deletePropertyByID(id) {
-  const data = await Property.findByIdAndDelete(id)
-  return {
-    data: data,
-    message: `deleted property successfully`,
-    status: true,
-    statusCode: 201,
-  };
+  try {
+    const data = await Inspection.find({
+      inspectionStatus: InspectionStatus.ACCEPTED,
+      propertyID: id
+    });
 
+    if (data.length > 0) {
+      return {
+        data: [],
+        message: "Unable To Delete Property",
+        status: false,
+        statusCode: 400,
+      };
+    }
+
+    const property = await Property.findByIdAndDelete(id);
+    if (property) {
+      return {
+        data: data,
+        message: "Property Deleted Successfully",
+        status: true,
+        statusCode: 200,
+      };
+    }
+    return {
+      data: [],
+      message: "Invalid Id",
+      status: false,
+      statusCode: 400,
+    };
+  } catch (error) {
+    return {
+      data: [],
+      message: error.message,
+      status: false,
+      statusCode: 400,
+    };
+  }
 }
 
 export {
